Fix ceiling return
Ceiling returns float in Python 2.7, but we need an int for range, so
cast it.

+++ wavedrom/waveform.py
@@ -169,7 +169,7 @@ class WaveDrom(SVGBase):
                 repeat += 1
             R.extend(self.gen_wave_brick(Top, This, stretch, repeat, subCycle))
 
-        for i in range(math.ceil(self.lane.phase)):
+        for i in range(int(math.ceil(self.lane.phase))):
             R = R[1:]
 
         return R
